app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ./app/requirements.in
#
aiohttp==3.9.1
    # via llama-index
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   openai
    #   sse-starlette
    #   starlette
apscheduler==3.10.4
    # via -r ./app/requirements.in
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
beautifulsoup4==4.12.2
    # via
    #   bs4
    #   llama-index
bs4==0.0.1
    # via -r ./app/requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via
    #   nltk
    #   uvicorn
cryptography==42.0.2
    # via
    #   msal
    #   pyjwt
dataclasses-json==0.6.3
    # via llama-index
deprecated==1.2.14
    # via llama-index
distro==1.9.0
    # via openai
dnspython==2.4.2
    # via pymongo
docx2txt==0.8
    # via -r ./app/requirements.in
exceptiongroup==1.1.3
    # via anyio
fastapi==0.104.1
    # via
    #   -r ./app/requirements.in
    #   sse-starlette
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.12.2
    # via llama-index
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
html2text==2020.1.16
    # via llama-hub
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via
    #   llama-index
    #   openai
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
joblib==1.3.2
    # via nltk
llama-hub==0.0.76
    # via -r ./app/requirements.in
llama-index==0.9.33
    # via
    #   -r ./app/requirements.in
    #   llama-hub
lxml==5.1.0
    # via python-pptx
marshmallow==3.20.2
    # via dataclasses-json
msal==1.26.0
    # via office365-rest-python-client
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.5.9
    # via llama-index
networkx==3.2.1
    # via llama-index
nltk==3.8.1
    # via llama-index
numpy==1.26.0
    # via
    #   llama-index
    #   pandas
office365-rest-python-client==2.5.5
    # via -r ./app/requirements.in
openai==1.8.0
    # via llama-index
packaging==23.2
    # via marshmallow
pandas==2.1.1
    # via llama-index
pillow==10.2.0
    # via python-pptx
pinecone-client==3.0.2
    # via -r ./app/requirements.in
psutil==5.9.8
    # via llama-hub
pyaml==23.12.0
    # via llama-hub
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via
    #   fastapi
    #   openai
pydantic-core==2.10.1
    # via pydantic
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
pymongo==4.5.0
    # via -r ./app/requirements.in
pypdf==3.17.0
    # via -r ./app/requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.1
    # via -r ./app/requirements.in
python-multipart==0.0.6
    # via -r ./app/requirements.in
python-pptx==0.6.23
    # via -r ./app/requirements.in
pytz==2023.3.post1
    # via
    #   apscheduler
    #   office365-rest-python-client
    #   pandas
pyyaml==6.0.1
    # via pyaml
regex==2023.8.8
    # via
    #   nltk
    #   tiktoken
requests==2.31.0
    # via
    #   llama-index
    #   msal
    #   office365-rest-python-client
    #   tiktoken
retrying==1.3.4
    # via llama-hub
six==1.16.0
    # via
    #   apscheduler
    #   python-dateutil
    #   retrying
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy[asyncio]==2.0.25
    # via
    #   llama-index
    #   sqlalchemy
sse-starlette==1.8.2
    # via -r ./app/requirements.in
starlette==0.27.0
    # via
    #   fastapi
    #   sse-starlette
tenacity==8.2.3
    # via llama-index
tiktoken==0.5.1
    # via llama-index
tqdm==4.66.1
    # via
    #   nltk
    #   openai
    #   pinecone-client
typing-extensions==4.8.0
    # via
    #   fastapi
    #   llama-index
    #   office365-rest-python-client
    #   openai
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index
tzdata==2023.3
    # via pandas
tzlocal==5.2
    # via apscheduler
urllib3==2.0.6
    # via
    #   pinecone-client
    #   requests
uvicorn==0.23.2
    # via
    #   -r ./app/requirements.in
    #   sse-starlette
wrapt==1.16.0
    # via deprecated
xlsxwriter==3.1.9
    # via python-pptx
yarl==1.9.2
    # via aiohttp
